requirements.txt
altgraph==0.17.3
certifi==2022.12.7
charset-normalizer==2.1.1
future==0.18.2
idna==3.4
keyboard==0.13.5
MouseInfo==0.1.3
mss==7.0.1
numpy==1.23.5
opencv-python==4.6.0.66
pefile==2022.5.30
Pillow==9.3.0
PyAutoGUI==0.9.53
PyDirectInput==1.0.4
PyGetWindow==0.0.9
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pywin32-ctypes==0.2.0
requests==2.28.1
urllib3==1.26.13